Refactor encoders
