+++ pkg/encoder/vpx/msvc.h
