+++ pkg/encoder/vpx/tools_common.c
