+++ pkg/encoder/vpx/tools_common.h
